feat: move `users` tasks to `database.users`
