+++ src/argilla/__main__.py
@@ -14,12 +14,11 @@
 #  limitations under the License.
 
 
-from argilla.tasks import database_app, server_app, training_app, users_app
+from argilla.tasks import database_app, server_app, training_app
 from argilla.tasks.async_typer import AsyncTyper
 
 app = AsyncTyper(rich_help_panel=True, help="Argilla CLI", no_args_is_help=True)
 
-app.add_typer(users_app, name="users")
 app.add_typer(database_app, name="database")
 app.add_typer(training_app, name="train")
 app.add_typer(server_app, name="server")
