+++ src/argilla/tasks/__init__.py
@@ -15,4 +15,3 @@
 from .database import app as database_app
 from .server import app as server_app
 from .training import app as training_app
-from .users import app as users_app
