+++ src/argilla/tasks/database/users/__main__.py
@@ -19,7 +19,9 @@ from .create_default import create_default
 from .migrate import migrate
 from .update import update
 
-app = AsyncTyper(help="Holds CLI commands for user and workspace management.", no_args_is_help=True)
+app = AsyncTyper(
+    help="CLI commands for user and workspace management using the database connection", no_args_is_help=True
+)
 
 app.command(name="create_default", help="Creates default users and workspaces in the Argilla database.")(create_default)
 app.command(name="create", help="Creates a user and add it to the Argilla database.", no_args_is_help=True)(create)
